watsonx/requirements.txt
langchain_community
langchain_text_splitters
langchain
ibm_watson_machine_learning
langchain
ibm_watsonx_ai
python-dotenv

opentelemetry-api
opentelemetry-distro
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-anthropic
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-aws-lambda
opentelemetry-instrumentation-chromadb
opentelemetry-instrumentation-cohere
opentelemetry-instrumentation-dbapi
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-jinja2
opentelemetry-instrumentation-langchain
opentelemetry-instrumentation-llamaindex
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-openai
opentelemetry-instrumentation-pinecone
opentelemetry-instrumentation-pymysql
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-sqlite3
opentelemetry-instrumentation-tortoiseorm
opentelemetry-instrumentation-transformers
opentelemetry-instrumentation-urllib
opentelemetry-instrumentation-urllib3
opentelemetry-instrumentation-wsgi
opentelemetry-propagator-aws-xray
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-semantic-conventions-ai
opentelemetry-util-http
traceloop-sdk
opentelemetry-instrumentation-watsonx
weaviate-client==3.26.2
chromadb
pysqlite3
pysqlite3-binary